fix: extract User type into lib/types.tsx
